Refactored loose auth ctrl methods into the AuthController class. Updated the auth routes to use new class methods.
