+++ app/routes/_auth.js
@@ -2,10 +2,7 @@ const createRouter = require("./utils/createRouter");
 
 const { login: loginVal } = require("./validation")
 
-const {
-    login,
-    authenticated
-} = require("../controllers/auth");
+const authCtrl = require("../controllers/auth");
 
 const loginCtlrConfig = {
     keyMap: { body: "credentials" }
@@ -17,7 +14,7 @@ module.exports = createRouter([
         post: {
             defaultError: "login",
             validation: loginVal,
-            ctrl: [ login, loginCtlrConfig ]
+            ctrl: [ authCtrl.binding.login, loginCtlrConfig ]
         }
     }],
 
@@ -25,7 +22,7 @@ module.exports = createRouter([
         post: {
             defaultError: "get the authenticated user",
             auth: true,
-            ctrl: authenticated
+            ctrl: authCtrl.binding.authenticated
         }
     }]
